refactor : package분리 및 인터페이스화.
