+++ src/main/java/step2/LottoRank.java
@@ -1,5 +1,7 @@
 package step2;
 
+import step2.vo.LottoResults;
+
 import java.util.List;
 
 public enum LottoRank {
