+++ src/main/java/step2/service/LottoGame.java
@@ -1,4 +1,7 @@
-package step2;
+package step2.service;
+
+import step2.vo.LottoResult;
+import step2.vo.LottoResults;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -8,20 +11,18 @@ public class LottoGame {
 
     private final int numOfLottoTicket;
 
-    private final DefaultLottoNumber defaultLottoNumber;
+    private final LottoNumber lottoNumber;
 
     public LottoGame(int numOfLottoTicket) {
         this.numOfLottoTicket = numOfLottoTicket;
-        this.defaultLottoNumber = new DefaultLottoNumber();
+        this.lottoNumber = new DefaultLottoNumber();
     }
 
     public LottoResults executeGame() {
         List<LottoResult> lottoResults = new ArrayList<>();
 
         for (int genNum = 0; genNum < numOfLottoTicket; genNum++) {
-            defaultLottoNumber.mixLottoNumbers();
-
-            List<Integer> extractedNumbers = defaultLottoNumber.extractLottoNumbers();
+            List<Integer> extractedNumbers = lottoNumber.extractLottoNumbers();
             Collections.sort(extractedNumbers);
             lottoResults.add(new LottoResult(extractedNumbers));
         }
