+++ src/main/java/step2/service/LottoNumber.java
@@ -1,0 +1,7 @@
+package step2.service;
+
+import java.util.List;
+
+public interface LottoNumber {
+    List<Integer> extractLottoNumbers();
+}
